refactor: add css theme variables
