+++ src/components/card-list/card-list.css
@@ -23,8 +23,9 @@
   line-height: 1;
   padding: 2rem 2rem;
   border: none;
-  background-color: #fff;
+  background-color: var(--white);
   width: 100%;
+  cursor: pointer;
 }
 
 .card-small--block {
