+++ src/components/footer/footer.css
@@ -1,6 +1,6 @@
 footer {
   height: 9.6rem;
-  background-color: #fdf2e9;
+  background-color: var(--accent-bg);
   display: grid;
   grid-template-columns: 1fr 2fr 1fr;
   align-items: center;
