+++ src/components/search-form/search-form.css
@@ -8,21 +8,21 @@
   padding: 1.2rem;
   font-size: 1.8rem;
   font-family: inherit;
-  color: inherit;
+  color: var(--text);
   border: none;
-  background-color: #fff;
+  background-color: var(--bg-placeholder);
   border-top-left-radius: 9px;
   border-bottom-left-radius: 9px;
   box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
 }
 
 .search-form input::placeholder {
-  color: #aaa;
+  color: var(--gray-placeholder);
 }
 
 .search-form input:focus {
   outline: none;
-  box-shadow: 0 0 0 0.4rem rgba(230, 125, 34, 0.5);
+  box-shadow: 0 0 0 0.4rem rgba(var(--shadow-focus), 0.5);
 }
 
 .search-form button {
