+++ src/context/context.ts
@@ -1,9 +1,8 @@
-import { ThemeContextProps, THEMES } from '@/interfaces/themes';
+import { ThemeContextProps } from '@/interfaces/themes';
 import { createContext, useContext } from 'react';
 
 export const ThemeContext = createContext<ThemeContextProps>({
   themeType: 'light',
-  theme: THEMES['light'],
   toggleTheme: () => {},
 });
