+++ src/index.css
@@ -3,15 +3,49 @@
   line-height: 1.5;
   font-weight: 400;
 
-  color: #213547;
-  background-color: #ffffff;
-
   font-synthesis: none;
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  --dark-body-bg: rgb(15 23 42 / 1);
-  --dark-div-bg: #1e293b;
+
+  --bg: #fff;
+  --bg-card: #fff;
+  --bg-placeholder: #fff;
+  --text: #213547;
+  --white: #fff;
+  --black: #000;
+
+  --accent: #e67e22;
+  --accent-bg: #fdf2e9;
+  --accent-hover: #cf711f;
+  --accent-hover-logo: #646cffaa;
+
+  --gray: #999;
+  --gray-title: #333;
+  --gray-placeholder: #aaa;
+  --shadow-focus: 230, 125, 34;
+  --shadow-link: 0, 0, 0;
+  --shadow-link-hover: 182, 53, 53;
+}
+
+[data-theme='dark'] {
+  --bg: #0f172a;
+  --bg-card: #1e293b;
+  --text: #94a3b8;
+  --accent: #262626;
+  --accent-bg: #374151;
+  --accent-hover: #171717;
+  --bg-placeholder: #1e293b;
+}
+
+[data-theme='light'] {
+  --bg: #fff;
+  --bg-card: #fff;
+  --text: #213547;
+  --accent: #e67e22;
+  --accent-bg: #fdf2e9;
+  --accent-hover: #cf711f;
+  --bg-placeholder: #fff;
 }
 
 * {
@@ -22,7 +56,7 @@
 
 *:focus {
   outline: none;
-  box-shadow: 0 0 0 0.4rem rgba(230, 125, 34, 0.5);
+  box-shadow: 0 0 0 0.4rem rgba(var(--shadow-focus), 0.5);
 }
 
 html {
@@ -34,6 +68,8 @@ html {
 body {
   min-width: 320px;
   height: 100vh;
+  color: var(--text);
+  background-color: var(--bg);
 }
 
 main {
@@ -47,10 +83,6 @@ a {
   width: auto;
   text-decoration: inherit;
   line-height: 0;
-}
-
-a:hover {
-  color: #747bff;
 }
 
 h1 {
@@ -83,13 +115,13 @@ h1 {
   font-family: inherit;
   color: #fff;
   border: none;
-  background-color: #e67e22;
+  background-color: var(--accent);
   transition: border-color 0.25s;
 }
 
 .btn:hover:not([disabled]),
 .btn:active:not([disabled]) {
-  background-color: #cf711f;
+  background-color: var(--accent-hover);
   cursor: pointer;
 }
 
@@ -104,17 +136,13 @@ h1 {
   transition: filter 300ms;
 }
 .logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
+  filter: drop-shadow(0 0 2em var(--accent-hover-logo));
 }
 
 @media (prefers-reduced-motion: no-preference) {
   a:nth-of-type(2) .logo {
     animation: logo-spin infinite 20s linear;
   }
-}
-
-.dark {
-  color-scheme: dark;
 }
 
 /* @media (prefers-color-scheme: light) {
